+++ batching_test.go
@@ -6,37 +6,37 @@ import (
 )
 
 func TestChannelClosing(t *testing.T) {
-	lines := make(chan string)
-	bufChan := make(chan *bytes.Buffer)
+	lines := make(chan *LogMessage)
+	bufChan := make(chan *LogMessage)
 
 	go Batch(lines, bufChan, 10)
-	lines <- "test log line"
+	lines <- &LogMessage{Lines: []byte("test log line")}
 	close(lines)
 
 	actual := <-bufChan
 	expected := "test log line\n"
-	if actual.String() != expected {
+	if string(actual.Lines) != expected {
 		t.Fatalf("expected \"%+v\", got \"%+v\"", expected, actual)
 	}
 }
 
 func TestBufferOverflow(t *testing.T) {
-	lines := make(chan string)
-	bufChan := make(chan *bytes.Buffer)
+	lines := make(chan *LogMessage)
+	bufChan := make(chan *LogMessage)
 
 	go Batch(lines, bufChan, 10)
 	filler := "test log line"
 	fillerLen := len(filler) + 1
 	for written := 0; written+fillerLen < 990000; written += fillerLen {
-		lines <- filler
+		lines <- &LogMessage{Lines: []byte(filler)}
 	}
-	lines <- "overflowed"
+	lines <- &LogMessage{Lines: []byte("overflowed")}
 	close(lines)
 
 	<-bufChan // throw away the big one
 	actual := <-bufChan
 	expected := "overflowed\n"
-	if actual.String() != expected {
+	if string(actual.Lines) != expected {
 		t.Fatalf("expected \"%+v\", got \"%+v\"", expected, actual)
 	}
 }
@@ -44,8 +44,8 @@ func TestBufferOverflow(t *testing.T) {
 // Batch()
 // Log lines larger than the max payload size (1 MB) should be dropped
 func TestBatchDropLogLine(t *testing.T) {
-	lines := make(chan string)
-	bufChan := make(chan *bytes.Buffer)
+	lines := make(chan *LogMessage)
+	bufChan := make(chan *LogMessage)
 
 	filler := "test log line"
 	buf := bytes.NewBuffer(make([]byte, 990000))
@@ -55,7 +55,7 @@ func TestBatchDropLogLine(t *testing.T) {
 	logline := buf.String()
 
 	go Batch(lines, bufChan, 10)
-	lines <- logline
+	lines <- &LogMessage{Lines: []byte(logline)}
 	close(lines)
 
 	// Nothing should be sent to bufChan since we are dropping message